+++ rector.php
@@ -2,25 +2,37 @@
 
 use Rector\CodeQuality\Rector\Concat\JoinStringConcatRector;
 use Rector\Config\RectorConfig;
-use Rector\PHPUnit\Rector\Class_\PreferPHPUnitSelfCallRector;
+use Rector\PHPUnit\CodeQuality\Rector\Class_\PreferPHPUnitSelfCallRector;
+use Rector\PHPUnit\CodeQuality\Rector\Class_\PreferPHPUnitThisCallRector;
+use Rector\PHPUnit\Set\PHPUnitSetList;
 use Rector\Set\ValueObject\SetList;
 
-return static function (RectorConfig $rectorConfig): void {
-    $rectorConfig->sets([
-        SetList::PHP_71,
-        SetList::PHP_72,
-        SetList::PHP_73,
-        SetList::PHP_74,
+return RectorConfig::configure()
+    ->withSets([
         SetList::CODE_QUALITY,
-    ]);
-    $rectorConfig->skip([
+        SetList::TYPE_DECLARATION,
+        SetList::RECTOR_PRESET,
+        PHPUnitSetList::PHPUNIT_40,
+        PHPUnitSetList::PHPUNIT_50,
+        PHPUnitSetList::PHPUNIT_60,
+        PHPUnitSetList::PHPUNIT_70,
+        PHPUnitSetList::PHPUNIT_80,
+        PHPUnitSetList::PHPUNIT_90,
+        PHPUnitSetList::PHPUNIT_100,
+        PHPUnitSetList::PHPUNIT_110,
+        PHPUnitSetList::PHPUNIT_CODE_QUALITY,
+    ])
+    ->withPhpSets()
+    ->withRules([PreferPHPUnitSelfCallRector::class])
+    ->withSkip([
+        PreferPHPUnitThisCallRector::class, // breaks tests
         JoinStringConcatRector::class => [
             __DIR__ . '/tests/CSVArrayTest.php', // keep `\r\n` for readability
         ],
-    ]);
-
-    $rectorConfig->rule(PreferPHPUnitSelfCallRector::class);
-
-    $rectorConfig->paths([__DIR__ . '/src', __DIR__ . '/tests']);
-    $rectorConfig->phpstanConfig(__DIR__ . '/phpstan.neon');
-};
+    ])
+    ->withPaths([__DIR__ . '/src', __DIR__ . '/tests'])
+    ->withBootstrapFiles([
+        // Rector uses PHPStan internally, which in turn requires Larastan to be set up correctly
+        __DIR__ . '/vendor/larastan/larastan/bootstrap.php',
+    ])
+    ->withPHPStanConfigs([__DIR__ . '/phpstan.neon']);