+++ CHANGELOG.md
@@ -9,6 +9,13 @@ See [GitHub releases](https://github.com/mll-lab/php-utils/releases).
 
 ## Unreleased
 
+## v5.6.0
+
+### Added
+
+- Improve code quality with automatic type declarations via Rector
+- Improve code quality with automatic rector rules for PHPUnit
+
 ## v5.5.2
 
 ### Fixed